requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Vector storage and ML
qdrant-client==1.7.0
sentence-transformers==2.2.2
numpy==1.24.3
torch>=2.0.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Graph processing
networkx==3.2
pyvis==0.3.2

# Web UI
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.3

# Utilities
loguru==0.7.2
python-multipart==0.0.6
httpx==0.25.2
tenacity==8.2.3

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
ruff==0.1.6
mypy==1.7.0
types-requests==2.31.0
pytest-cov==4.1.0